refactor: `Validate`を移動し、`Note`と`Score`でも利用
