+++ crates/voicevox_core/src/engine/talk/audio_query/validated.rs
@@ -5,7 +5,6 @@ use std::{
 };
 
 use duplicate::duplicate_item;
-use serde::de::DeserializeOwned;
 use tracing::warn;
 
 use crate::error::{InvalidQueryError, InvalidQueryErrorSource};
@@ -14,6 +13,7 @@ use super::{
     super::super::{
         acoustic_feature_extractor::{Consonant, NonConsonant},
         sampling_rate::SamplingRate,
+        validate::Validate as _,
         DEFAULT_SAMPLING_RATE,
     },
     AccentPhrase, AudioQuery, Mora,
@@ -132,30 +132,6 @@ impl AudioQuery {
     // TODO: この層を破壊
     pub(crate) fn to_validated(&self) -> crate::Result<ValidatedAudioQuery<'_>> {
         ValidatedAudioQuery::new(self).map_err(Into::into)
-    }
-}
-
-pub trait Validate: DeserializeOwned {
-    const NAME: &str;
-    fn validate(&self) -> crate::Result<()>;
-
-    fn validation_error_description() -> String {
-        format!("不正な{}です", Self::NAME)
-    }
-}
-
-#[duplicate_item(
-    T S validation;
-    [ AudioQuery ] [ "AudioQuery" ] [ Self::validate ];
-    [ AccentPhrase ] [ "アクセント句" ] [ Self::validate ];
-    [ Mora ] [ "モーラ" ] [ Self::validate ];
-    [ Vec<AccentPhrase> ] [ "アクセント句の列" ] [ |this: &Self| this.iter().try_for_each(AccentPhrase::validate) ];
-)]
-impl Validate for T {
-    const NAME: &str = S;
-
-    fn validate(&self) -> crate::Result<()> {
-        (validation)(self)
     }
 }
